14일 오후 수정

+++ job1/jobs/views.py
@@ -1231,10 +1231,10 @@ def JB103_4(request): # 직무 현황표, 기술서 print
 
         # pymysql을 사용하여 데이터베이스에 연결
         conn = pymysql.connect(
-            host='130.1.200.200', # 데이터베이스 주소
+            host='130.1.112.100', # 데이터베이스 주소
             user='cdh', # 데이터베이스 사용자 이름
-            password='1234', # 데이터베이스 비밀번호
-            db='jobdb',
+            password='cdh0706**', # 데이터베이스 비밀번호
+            db='betadb',
             charset='utf8',
             cursorclass=pymysql.cursors.DictCursor
         )
@@ -3134,7 +3134,7 @@ def BS300_6(request): # 조직 그룹 탭에서 저장, 취소
             deptGrpData = request.POST.get('deptGrpData')
             data2 = json.loads(deptGrpData)
             df_dept = pd.DataFrame(data2) # 그룹별 부서 테이블과 비교
-            print(df_dept)
+            # print(df_dept)
 
             # 먼저 BsDeptGrpDomain 테이블을 df_domain_grp와 비교하여 업데이트. JB103_3과 비슷한 로직
             # 기존 데이터(해당 회기의 데이터)를 가져와서 비교한다.
@@ -3164,21 +3164,35 @@ def BS300_6(request): # 조직 그룹 탭에서 저장, 취소
             data_list2 = [{'dept_domain': rows.dept_domain_id, 'dept_grp_nm' : rows.dept_grp_nm_id, 'dept_cd':rows.dept_cd_id, 'dept_seq': rows.dept_seq } for rows in original_rows2]
             df_2 = pd.DataFrame(data_list2)
 
-            # 비교 하는 부분 - merge 기능을 이용해 추가된 행, 삭제된 행을 추출할 것이다. 수정은 삭제 후 추가로 볼 것이다.
-            # df_left_dept : df2(DB)에 있고 df_dept(UI)에 없는 것. 즉, 삭제된 것
-            df_left_dept = pd.merge(df_2, df_dept, how='outer', indicator=True).query('_merge == "left_only"').drop(columns=['_merge']).reset_index(drop=True)
-            # df_right_dept : df_dept(UI)에 있고 df2(DB)에 없는 것. 즉, 추가된 것
-            df_right_dept = pd.merge(df_2, df_dept, how='outer', indicator=True).query('_merge == "right_only"').drop(columns=['_merge']).reset_index(drop=True)
-            print(df_right_dept)
+            # df_2가 비어있지 않은 경우(BsDeptGrp에 데이터 있는 경우)에만 비교를 진행한다.
+            if not df_2.empty:
+                
+                if not df_dept.empty: # df_dept가 비어있지 않은 경우, 즉 화면에서 어느 한 부서라도 체크를 한 경우에는 비교를 진행한다.
 
-            # df_left_dept부터 다룬다. df_left_dept에 있는 것들은 삭제해야 하는 것들이다. DB에 접근해서 지워줌
-            for i in range(0, len(df_left_dept)):
-                BsDeptGrp.objects.filter(prd_cd_id=prd_cd_selected, dept_domain_id=df_left_dept.iloc[i, 0], dept_grp_nm_id=df_left_dept.iloc[i, 1], dept_cd_id=df_left_dept.iloc[i, 2]).delete()
+                    # 비교 하는 부분 - merge 기능을 이용해 추가된 행, 삭제된 행을 추출할 것이다. 수정은 삭제 후 추가로 볼 것이다.
+                    # df_left_dept : df2(DB)에 있고 df_dept(UI)에 없는 것. 즉, 삭제된 것
+                    df_left_dept = pd.merge(df_2, df_dept, how='outer', indicator=True).query('_merge == "left_only"').drop(columns=['_merge']).reset_index(drop=True)
+                    # df_right_dept : df_dept(UI)에 있고 df2(DB)에 없는 것. 즉, 추가된 것
+                    df_right_dept = pd.merge(df_2, df_dept, how='outer', indicator=True).query('_merge == "right_only"').drop(columns=['_merge']).reset_index(drop=True)
+                    print(df_right_dept)
 
-            # df_right_dept 다룬다. df_right_dept에 있는 것들은 추가해야 하는 것들이다. DB에 접근해서 추가해줌
-            for i in range(0, len(df_right_dept)):
-                BsDeptGrp.objects.create(prd_cd_id=prd_cd_selected, dept_domain_id=df_right_dept.iloc[i, 0], dept_grp_nm_id=df_right_dept.iloc[i, 1], dept_cd_id=df_right_dept.iloc[i, 2],
-                                          dept_seq=df_right_dept.iloc[i, 3])
+                    # df_left_dept부터 다룬다. df_left_dept에 있는 것들은 삭제해야 하는 것들이다. DB에 접근해서 지워줌
+                    for i in range(0, len(df_left_dept)):
+                        BsDeptGrp.objects.filter(prd_cd_id=prd_cd_selected, dept_domain_id=df_left_dept.iloc[i, 0], dept_grp_nm_id=df_left_dept.iloc[i, 1], dept_cd_id=df_left_dept.iloc[i, 2]).delete()
+
+                    # df_right_dept 다룬다. df_right_dept에 있는 것들은 추가해야 하는 것들이다. DB에 접근해서 추가해줌
+                    for i in range(0, len(df_right_dept)):
+                        BsDeptGrp.objects.create(prd_cd_id=prd_cd_selected, dept_domain_id=df_right_dept.iloc[i, 0], dept_grp_nm_id=df_right_dept.iloc[i, 1], dept_cd_id=df_right_dept.iloc[i, 2],
+                                                dept_seq=df_right_dept.iloc[i, 3])
+                        
+                else: # df_dept가 비어있는 경우, 즉 화면에서 어느 한 부서도 체크를 하지 않은 경우에는 BsDeptGrp 테이블을 다 없애준다.
+                    BsDeptGrp.objects.filter(prd_cd_id=prd_cd_selected).delete()
+
+            # df2가 비어있는 경우, 즉 BsDeptGrp가 비어있는 경우는 df_dept 이용해서 BsDeptGrp 테이블에 추가해준다.
+            else:
+                for i in range(0, len(df_dept)):
+                    BsDeptGrp.objects.create(prd_cd_id=prd_cd_selected, dept_domain_id=df_dept.iloc[i, 0], dept_grp_nm_id=df_dept.iloc[i, 1], dept_cd_id=df_dept.iloc[i, 2],
+                                            dept_seq=df_dept.iloc[i, 3])
 
             # 편집을 다 끝내고 나면 다시 화면을 띄워준다.
             # BsDeptGrp를 가져와서 json으로 변경시켜준다. JB103이랑 비슷하게 한다.
@@ -3193,7 +3207,7 @@ def BS300_6(request): # 조직 그룹 탭에서 저장, 취소
 
             original_rows2 = BsDeptGrpDomain.objects.filter(prd_cd_id=prd_cd_selected)
             data_list2 = [{'dept_domain': rows.dept_domain, 'dept_grp_nm' : rows.dept_grp_nm, 'domain_seq': rows.domain_seq,
-                           'grp_seq': rows.grp_seq } for rows in original_rows2]
+                        'grp_seq': rows.grp_seq } for rows in original_rows2]
             df2 = pd.DataFrame(data_list2).sort_values(by=['domain_seq', 'grp_seq']).reset_index(drop=True)
             df_json2 = df2.to_json(orient='records')
             # 엑셀 저장
@@ -3209,6 +3223,7 @@ def BS300_6(request): # 조직 그룹 탭에서 저장, 취소
             context.update({'data_bs_dept_grp' : df_json})
             context.update({'data_bs_dept_grp_domain' : df_json2})
             context.update({'data_bs_dept' : df_json3})
+           
 
         if action == 'action2': # 취소 버튼 눌렀을 때
 
@@ -7230,6 +7245,67 @@ def JB109_2(request): # 업무량 분석화면 - 탭 선택 후 선택한 탭을
                 'dept_mgr_yn' : get_dept_mgr_yn(request.user.username),
             }
 
+        if span_name == 'span3': # 업무량 구성비 탭 선택했을 때
+
+            context = {
+                'prd_list' : BsPrd.objects.all(),
+                'title' : '업무량 분석', # 제목
+                'prd_cd_selected' : prd_cd_selected,
+                'tab' : 'tab3',
+                'dept_mgr_yn' : get_dept_mgr_yn(request.user.username),
+            }
+
+            # # 바로 dataframe 만들어준다.
+            # # BsDeptGrpDomain 테이블로부터 dept_domain, dept_grp_nm, domain_seq, grp_seq를 가져온다.
+            # # BsDeptGrp 테이블로부터 dept_domain, dept_grp_nm, dept_cd, dept_seq를 가져온다.
+            # # 가져와서 dataframe 각각 만들어주고 merge한다.
+            # target = BsDeptGrpDomain.objects.filter(prd_cd_id=prd_cd_selected)
+            # data_list = [{'dept_domain' : rows.dept_domain, 'dept_grp_nm' : rows.dept_grp_nm, 'domain_seq': rows.domain_seq,
+            #             'grp_seq': rows.grp_seq} for rows in target]
+            # df1 = pd.DataFrame(data_list)
+            # target2 = BsDeptGrp.objects.filter(prd_cd_id=prd_cd_selected)
+            # data_list2 = [{'dept_domain' : rows.dept_domain_id, 'dept_grp_nm' : rows.dept_grp_nm_id, 'dept_cd': rows.dept_cd_id,
+            #             'dept_seq': rows.dept_seq} for rows in target2]
+            # df2 = pd.DataFrame(data_list2)
+            # df3 = pd.merge(df1, df2)
+
+            # # df3에 dept_po 열을 추가한다. dept_po 데이터는 BsDept 테이블에서 해당 부서코드의 dept_po 값을 가져온다.
+            # df3['dept_po'] = df3['dept_cd'].apply(lambda x: BsDept.objects.get(prd_cd=prd_cd_selected, dept_cd=x).dept_po)
+
+            # # df3에 prfrm_tm_ann 열을 추가한다. df3의 각 행에 대하여(for문 활용), 그 행으 dept_cd에 해당하는 JobTask 테이블의 prfrm_tm_ann 값을 가져온다.
+            # # JobTask object들의 prfrm_tm_ann 값을 더해주면 그 행의 prfrm_tm_ann이다.
+            # for i in range(len(df3)):
+
+            #     sum_prfrm_tm_ann = 0
+
+            #     sum_target = JobTask.objects.filter(prd_cd_id=prd_cd_selected, dept_cd_id=df3.iloc[i]['dept_cd'])
+
+            #     for rows in sum_target:
+            #         if rows.prfrm_tm_ann == None:
+            #             rows.prfrm_tm_ann = 0
+            #         sum_prfrm_tm_ann = sum_prfrm_tm_ann + rows.prfrm_tm_ann
+
+            #     df3.loc[i, 'prfrm_tm_ann'] = float(sum_prfrm_tm_ann)
+
+            # #df3에 po_nec 열을 추가한다. df3의 prfrm_tm_ann열을 BsStdWrkTm 테이블의 std_wrk_tm으로 나눈 값을 넣어준다.
+            # std_wrk_tm = float(BsStdWrkTm.objects.get(prd_cd=prd_cd_selected).std_wrk_tm)
+
+            # df3['po_nec'] = round(df3['prfrm_tm_ann']/std_wrk_tm, 1)
+
+            # print(df3)
+
+            # # df3에 환산 업무량 열을 추가한다. df3의 각 행에 대해서 계산해준다. for문 활용한다.
+            # for i in range(len(df3)):
+            #     task_list = JobTask.objects.filter(prd_cd_id=prd_cd_selected, dept_cd_id=df3.iloc[i]['dept_cd'])
+            #     sum_tm_cal = 0
+
+            #     # task_list의 work_grade를 보고, 각 work_grade 에 따른 업무량 가중치를 .....
+
+            
+
+
+
+
     return render(request, 'jobs/JB109.html', context)
 
 
@@ -7697,10 +7773,10 @@ def BsMbrArrange(prd, dept): # 부서원 표시 함수 - 수정해야함
 def copy_period_data(period_old, period_new):
     # 데이터베이스 연결 파라미터
     user_id = 'cdh'  # 사용자 이름
-    pwd = '1234'  # 비밀번호
-    db_host = '130.1.200.200'  # 호스트명/IP
+    pwd = 'cdh0706**'  # 비밀번호
+    db_host = '130.1.112.100'  # 호스트명/IP
     db_port = 3306  # 포트번호 (고정값)
-    db_name = "jobdb"  # 사용할 데이터베이스 jobdb
+    db_name = "betadb"  # 사용할 데이터베이스 betadb
 
     dict_table = {  # 테이블 목록
         'bs_prd': '회기',
@@ -7772,10 +7848,10 @@ def copy_period_data(period_old, period_new):
 def delete_period_data(period):
     # 데이터베이스 연결 파라미터
     user_id = 'cdh'  # 사용자 이름
-    pwd = '1234'  # 비밀번호
-    db_host = '130.1.200.200'  # 호스트명/IP
+    pwd = 'cdh0706**'  # 비밀번호
+    db_host = '130.1.112.100'  # 호스트명/IP
     db_port = 3306  # 포트번호 (고정값)
-    db_name = "jobdb"  # 사용할 데이터베이스 jobdb
+    db_name = "betadb"  # 사용할 데이터베이스 betadb
 
     dict_table = {  # 테이블 목록
         'job_spcfc': '직무명세서',
